Add search name algo and bring search and page logic to controller
