+++ Areas/Admin/Services/Practitioner/PractitionerService.cs
@@ -60,14 +60,23 @@ namespace PainAssessment.Areas.Admin.Services
     }
 }
 
-public static class ListExtensions
-{
-    public static List<List<T>> ChunkBy<T>(this List<T> source, int chunkSize)
-    {
-        return source
-            .Select((x, i) => new { Index = i, Value = x })
-            .GroupBy(x => x.Index / chunkSize)
-            .Select(x => x.Select(v => v.Value).ToList())
-            .ToList();
-    }
-}
+//public static class ListExtensions
+//{
+//    public static List<List<T>> ChunkBy<T>(this List<T> source, int chunkSize)
+//    {
+//        return source
+//            .Select((x, i) => new { Index = i, Value = x })
+//            .GroupBy(x => x.Index / chunkSize)
+//            .Select(x => x.Select(v => v.Value).ToList())
+//            .ToList();
+//    }
+//}
+
+//public static class IEnumerableExtensions
+//{
+//    public static IEnumerable<IEnumerable<T>> ChunkBy<T>(this IEnumerable<T> source, int chunkSize)
+//    {
+//        for (int i = 0; i < source.Count(); i += chunkSize)
+//            yield return source.Skip(i).Take(chunkSize);
+//    }
+//}
