+++ Areas/Admin/Views/Practitioners/Index.cshtml
@@ -24,16 +24,19 @@
     </div>
 
     <div class="d-flex-col mt-2">
-        <div class="d-flex justify-content-end">
+
+        <form class="d-flex justify-content-end" method="get" asp-action="Index">
             <input class="form-control me-2"
                    type="search"
+                   name="name"
                    placeholder="Search with Name"
                    aria-label="Search"
-                   id="" />
-            <button class="btn btn-outline-success">
+                   value="@ViewData["name"]" />
+            <button type="submit" class="btn btn-outline-success">
                 Search
             </button>
-        </div>
+        </form>
+
 
         <p class="d-flex justify-content-end text-muted">Displaying @((((int)ViewData["current_page"]-1)*8)+1) to @((((int)ViewData["current_page"]-1)*8)+Model.Count())  out of @ViewData["total_count"] records</p>
 
@@ -102,7 +105,7 @@
                         {
                             <li class="page-item">
                                 <a class="page-link"
-                                   aria-label="Previous" asp-action="Index" asp-route-page=@((int) ViewData["current_page"]-1)>
+                                   aria-label="Previous" asp-action="Index" asp-route-page=@((int) ViewData["current_page"]-1) asp-route-name=@ViewData["name"]>
                                     <span aria-hidden="true">&laquo;</span>
                                 </a>
                             </li>
@@ -122,7 +125,7 @@
                             {
 
                                 <li class="page-item">
-                                    <a class="page-link" asp-action="Index" asp-route-page=@i>@i</a>
+                                    <a class="page-link" asp-action="Index" asp-route-page=@i asp-route-name=@ViewData["name"]>@i</a>
                                 </li>
                             }
                         }
@@ -131,7 +134,7 @@
                         {
                             <li class="page-item">
                                 <a class="page-link"
-                                   aria-label="Next" asp-action="Index" asp-route-page=@((int)ViewData["current_page"] + 1)>
+                                   aria-label="Next" asp-action="Index" asp-route-page=@((int)ViewData["current_page"] + 1) asp-route-name=@ViewData["name"]>
                                     <span aria-hidden="true">&raquo;</span>
                                 </a>
                             </li>
